Create GameOfLife.pde
This version is much better than my lengthy OOP one. About 2/3s of the way into writing it, i realized that it would be much simpler with just an array of the state values. I tried a 1D array but eventually settled on a 2D since it would be a little simpler.